TD-1868 Improve memory management in epub export

+++ pathway/epubConvert/Exportepub.cs
@@ -365,20 +365,29 @@ namespace SIL.PublishingSolution
                 // customize the CSS file based on the settings
                 CustomizeCSS(mergedCSS);
 
-                foreach (string file in splitFiles)
+                string getPsApplicationPath = Common.GetPSApplicationPath();
+                string xsltProcessExe = Path.Combine(getPsApplicationPath, "XslProcess.exe");
+                inProcess.SetStatus("Apply Xslt Process in html file");
+                if (File.Exists(xsltProcessExe))
                 {
-                    inProcess.SetStatus("Converting file to .epub format:" + file);
-                    if (File.Exists(file))
+                    foreach (string file in splitFiles)
                     {
-                        Common.XsltProcess(file, xsltFullName, "_.xhtml");
-                        // add this file to the html files list
-                        htmlFiles.Add(Path.Combine(Path.GetDirectoryName(file), (Path.GetFileNameWithoutExtension(file) + "_.xhtml")));
-                        // clean up the un-transformed file
-                        File.Delete(file);
-                    }
-                    inProcess.PerformStep();
-                }
+                        if (File.Exists(file))
+                        {
+                            Common.RunCommand(xsltProcessExe,
+                                              string.Format("{0} {1} {2} {3}", file, xsltFullName, "_.xhtml",
+                                                            getPsApplicationPath), 1);
 
+                            //Common.XsltProcess(file, xsltFullName, "_.xhtml");
+                            // add this file to the html files list
+                            htmlFiles.Add(Path.Combine(Path.GetDirectoryName(file),
+                                                       (Path.GetFileNameWithoutExtension(file) + "_.xhtml")));
+                            // clean up the un-transformed file
+                            File.Delete(file);
+                        }
+                    }
+                }
+                inProcess.PerformStep();
                 // create the "epub" directory structure and copy over the boilerplate files
                 inProcess.SetStatus("Creating .epub structure");
                 sb.Append(tempFolder);
